COMP: Remove compiler warnings.

+++ BRAINSCut/cnpy/cnpy.h
@@ -46,7 +46,7 @@ namespace cnpy {
 
     template<typename T> std::vector<char>& operator+=(std::vector<char>& lhs, const T rhs) {
         //write in little endian
-        for(char byte = 0; byte < sizeof(T); byte++) {
+        for(char byte = 0; static_cast<size_t>(byte) < sizeof(T); byte++) {
             char val = *((char*)&rhs+byte); 
             lhs.push_back(val);
         }
